transferred LinkButton to components
